plrctrls,options: introduce Gameplay.noAutowalkOnSecondaryAction option
This commit introduces the `Gameplay.noAutowalkOnSecondaryAction`
option in the `diablo.ini` file, which disables auto-walk when
pressing the secondary action:

  Gameplay.noAutowalkOnSecondaryAction=1

The default behavior (auto-walk on secondary action) disrupts
automation (bot) when a player stands near an item/barrel and a
trigger (warp to a town or next level). An attempt to break a barrel
or pick up an item (pressing the secondary action) will cause a warp
to a town/next level. Unfortunately, this is unavoidable even when the
player is oriented towards the item/barrel - the game warps anyway.

This commit makes it possible to disable this default behavior, which
is generally considered useless for automation (bot) or AI agents.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>