plrctrls: disable monster auto-pursuing on prime action
Once the primary action is pressed on a controller/pad,
the hero starts pursuing a monster. If there are a lot
of monsters around, sequential press of the primary
action buttuin makes the hero run around across the
whole dungeon in absolutely autonomous manner.

Patch sets amount of max steps to 0, disabling the
auto-pursuing, which should be good to teach AI
where to move the hero and not always rely on the
primary action button.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>