plrctrls,options: introduce Gameplay.noMonsterAutoPursuing option
This commit introduces the `Gameplay.noMonsterAutoPursuing` option in
the `diablo.ini` file, which disables monster auto-pursuing on primary
actions if set to `1`:

  Gameplay.noMonsterAutoPursuing=1

Once the primary action is pressed on a controller/pad, the hero
begins pursuing a monster. If there are a lot of monsters around,
repeatedly pressing the primary action button makes the hero run
around the entire dungeon in a completely autonomous manner.

The patch sets the amount of max steps to 0, disabling auto-pursuing.
This should be useful for teaching the AI where to move the hero
instead of always relying on the primary action button.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>