diablo,options: introduce Gameplay.gameTicksPerStep and stepMode options
Introduce new options: `Gameplay.gameTicksPerStep` and
`Gameplay.stepMode`, which allow Diablo to run in a step mode and skip
a few game ticks per step.

- The option `gameTicksPerStep` specifies the number of game ticks
that will be executed in a single step.

- The option `stepMode` enables step-mode, meaning the game loop does
not run freely but waits for further events from an agent

The purpose of the step mode with a few game ticks per step (usually
10) is for AI training.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>