diablo,options: introduce Gameplay.gameTicksPerStep option
Introduce a new option: `Gameplay.gameTicksPerStep`, which allows
Diablo to run in a single-step mode. The option itself specifies the
number of game ticks that will be executed in a single step. If
`Gameplay.gameTicksPerStep` is 0, run-time mode (default Diablo mode)
is used.

The purpose of the step mode is AI training.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>