diablo: split EventHandler on actual handler and poller callbacks
There are various message handlers in the code, including:

  PollEvent()
  FetchMessage()
  FetchMessage_Real()

These functions internally call SDL_PollEvent() and then handle the
received events differently depending on the context.

This commit separates the actual handlers from the poll invocations by
adding an additional `poll` callback parameter. This parameter can be
either SDL_PollEvent() or something else, allowing the caller of the
message handler to decide which events need to be processed.

This commit does not change the logic of event processing. Its main
purpose is to make it possible to replace SDL_PollEvent() with
SDL_PeepEvents() in the next patch.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>