diablo: do not drop other events in the `ShowProgress` routine
`ShowProgress` is called on any custom event, such as when a player
warps to a new level. It is designed to handle only custom events and
discards others from the queue. This behavior results in lost
keystrokes sent by the AI agent. For instance, the issue was
reproduced when attempting to pause the game by sending the PAUSE key,
which was lost due to the behavior of `ShowProgress`.

This change introduces a poller along with the custom handler. For the
`ShowProgress` routine, the poller will peek at only custom events,
leaving others in the queue.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>