lighting: rename variables and add explicit comments for clarity
This patch improves clarity and readability without affecting
functionality:

1. Renames `VisionCrawlTable` to `VisionRays` and `crawl` to
`rayPoint` for better clarity on the purpose of these structures.

2. Renames `factors` to `quadrants` to reflect the actual purpose of
the mirror operation along the X or Y coordinate lines.

3. Replaces C-style 2D arrays of rays with C++ vectors to enable the
use of variable lengths for rays, allowing zero points at the end of
each ray to be trimmed.

4. Removes the `RadiusAdj` array, whose purpose was to reduce the ray
size to the number of non-zero points and make circle of rays look
more accurate. Since the C-style array of rays is replaced with a
vector definition, `RadiusAdj` becomes useless.

5. Adds more explicit comments to simplify the understanding of the
ray casting algorithm.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>