CMake: introduce compilation test for the linker symbols
The patch includes a CMake compilation test that verifies whether the
resulting binary includes the `__bss_start` and `_end` linker
symbols. Should be the case for the GNU linker. If the binary contains
these symbols, the `HAVE_LINKER_BSS_SYMBOLS` macro will be defined.

In the next commit, the `HAVE_LINKER_BSS_SYMBOLS` macro will be used
to share the .bss and .data sections with the external
application. Stay tuned.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>