gmenu: return an indication of handled keys from `gmenu_presskeys()`
This patch updates the function `gmenu_presskeys()` to return `true`
if the keys are actually handled. This allows the caller function
`PressKey()` to continue processing keys and not silently drop them,
even when the game menu is active.

Why bother? First of all, there should not be any reasons not to
handle keys once the game menu is open. For example, why not handle
the "Quick Load" shortcut once the player is dead and the menu is
shown on the screen?

The other reason is more practical. In headless mode, once the game is
controlled by a 3rd party application (training of an AI agent, for
example), lost key presses lead to a broken training procedure.
Therefore, even if the menu is raised, it is crucial not to lose a
keystroke.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>