diablo: introduce a flag to skip hero menu if new level generated
AI sends a KEY_NEW key, which triggers the generation and start of a
new level. This patch introduces a flag that helps to skip all hero
selection menus if the NEW key was sent by the AI.

The purpose of this flag is to automate level generation as much as
possible.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>