diablo: don't forget to call `ClearUniqueItemFlags()` for headless mode
By mistake, the `ClearUniqueItemFlags()` function was not called for
headless mode, which leaves item flags initialized from the previous
game run. This is especially important when the seed is initialized:
if flags are not cleared, `TryRandomUniqueItem()` follows a different
path on a second run, causing the random generator to diverge.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>